+++ mex/common/organigram/extract.py
@@ -53,6 +53,36 @@ def get_unit_synonyms(extracted_unit: ExtractedOrganizationalUnit) -> list[str]:
     )
 
 
+def get_extracted_unit_by_synonyms(
+    extracted_units: Iterable[ExtractedOrganizationalUnit],
+) -> dict[str, ExtractedOrganizationalUnit]:
+    """Return a mapping from unit alt_label and label to their organizational units.
+
+    There will be multiple entries per unit mapping to the same organizational unit.
+
+    Args:
+        extracted_units: Iterable of extracted units
+
+    Raises:
+        MExError: If the same entry maps to different organizational units
+
+    Returns:
+        Mapping from unit synonyms to extracted units
+    """
+    synonym_dict: dict[str, ExtractedOrganizationalUnit] = {}
+    for extracted_unit in extracted_units:
+        for synonym in get_unit_synonyms(extracted_unit):
+            if synonym in synonym_dict and synonym_dict[synonym] != extracted_unit:
+                msg = (
+                    f"Conflict: label '{synonym}' is associated with merged unit IDs "
+                    f"{synonym_dict[synonym].stableTargetId} and "
+                    f"{extracted_unit.stableTargetId}."
+                )
+                raise MExError(msg)
+            synonym_dict[synonym] = extracted_unit
+    return synonym_dict
+
+
 def get_unit_merged_ids_by_synonyms(
     extracted_units: Iterable[ExtractedOrganizationalUnit],
 ) -> dict[str, MergedOrganizationalUnitIdentifier]:
@@ -63,26 +93,13 @@ def get_unit_merged_ids_by_synonyms(
     Args:
         extracted_units: Iterable of extracted units
 
-    Raises:
-        MExError: If the same entry maps to different merged IDs
-
     Returns:
         Mapping from unit synonyms to stableTargetIds
     """
-    synonym_dict: dict[str, MergedOrganizationalUnitIdentifier] = {}
-    for extracted_unit in extracted_units:
-        for synonym in get_unit_synonyms(extracted_unit):
-            if (
-                synonym in synonym_dict
-                and synonym_dict[synonym] != extracted_unit.stableTargetId
-            ):
-                msg = (
-                    f"Conflict: label '{synonym}' is associated with merged unit IDs "
-                    f"{synonym_dict[synonym]} and {extracted_unit.stableTargetId}."
-                )
-                raise MExError(msg)
-            synonym_dict[synonym] = extracted_unit.stableTargetId
-    return synonym_dict
+    return {
+        key: value.stableTargetId
+        for key, value in get_extracted_unit_by_synonyms(extracted_units).items()
+    }
 
 
 def get_unit_merged_ids_by_emails(